some cleanup of tests
