+++ api-test/ArtefactTest.cs
@@ -62,7 +62,7 @@ namespace SQE.ApiTest
         {
             // Arrange
             var allArtefacts = (await GetRandomEditionArtefacts()).artefacts; // Find edition with artefacts
-            var newEdition = await HttpRequest.CreateNewEdition(_client, allArtefacts.First().editionId); // Clone it
+            var newEdition = await EditionHelpers.CreateCopyOfEdition(_client, allArtefacts.First().editionId); // Clone it
 
             const string masterImageSQL = "SELECT sqe_image_id FROM SQE_image WHERE type = 0 ORDER BY RAND() LIMIT 1";
             var masterImageId = await _db.RunQuerySingleAsync<uint>(masterImageSQL, null);
@@ -152,7 +152,7 @@ namespace SQE.ApiTest
             // Arrange
             var allArtefacts = (await GetRandomEditionArtefacts()).artefacts; // Find edition with artefacts
             var artefact = allArtefacts.First();
-            var newEdition = await HttpRequest.CreateNewEdition(_client, artefact.editionId); // Clone it
+            var newEdition = await EditionHelpers.CreateCopyOfEdition(_client, artefact.editionId); // Clone it
             
             // Act
             var (response, writtenArtefact) = await HttpRequest.SendAsync<string, string>(_client, HttpMethod.Delete,
@@ -169,7 +169,7 @@ namespace SQE.ApiTest
                 await HttpRequest.GetJWTAsync(_client));
             Assert.Equal(HttpStatusCode.NotFound, delResponse.StatusCode);
 
-            await HttpRequest.DeleteEdition(_client, newEdition, true);
+            await EditionHelpers.DeleteEdition(_client, newEdition, true);
         }
         
         /// <summary>
@@ -182,7 +182,7 @@ namespace SQE.ApiTest
             // Arrange
             var allArtefacts = (await GetRandomEditionArtefacts()).artefacts; // Find edition with artefacts
             var artefact = allArtefacts.First();
-            var newEdition = await HttpRequest.CreateNewEdition(_client, artefact.editionId); // Clone it
+            var newEdition = await EditionHelpers.CreateCopyOfEdition(_client, artefact.editionId); // Clone it
             var newArtefactName = _faker.Random.Words(5);
             var newArtefactPosition = RandomPosition();
             const string newArtefactShape = "POLYGON((0 0,0 200,200 200,0 200,0 0),(5 5,5 25,25 25,25 5,5 5),(77 80,77 92,102 92,102 80,77 80))";
@@ -254,7 +254,7 @@ namespace SQE.ApiTest
             Assert.Equal(otherTransform, updatedAllArtefact.mask.transformMatrix);
             Assert.Equal(artefact.name, updatedAllArtefact.name);
 
-            await HttpRequest.DeleteEdition(_client, newEdition, true);
+            await EditionHelpers.DeleteEdition(_client, newEdition, true);
         }
 
         #endregion Access artefacts (should succeed)
@@ -269,7 +269,7 @@ namespace SQE.ApiTest
         {
             // Arrange
             var allArtefacts = (await GetRandomEditionArtefacts()).artefacts; // Find edition with artefacts
-            var newEdition = await HttpRequest.CreateNewEdition(_client, allArtefacts.First().editionId); // Clone it
+            var newEdition = await EditionHelpers.CreateCopyOfEdition(_client, allArtefacts.First().editionId); // Clone it
 
             const string masterImageSQL = "SELECT sqe_image_id FROM SQE_image WHERE type = 0 ORDER BY RAND() LIMIT 1";
             var masterImageId = await _db.RunQuerySingleAsync<uint>(masterImageSQL, null);
@@ -291,7 +291,7 @@ namespace SQE.ApiTest
             // Assert
             Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
 
-            await HttpRequest.DeleteEdition(_client, newEdition, true);
+            await EditionHelpers.DeleteEdition(_client, newEdition, true);
         }
         
         /// <summary>
@@ -304,7 +304,7 @@ namespace SQE.ApiTest
             // Arrange
             var allArtefacts = (await GetRandomEditionArtefacts()).artefacts; // Find edition with artefacts
             var artefact = allArtefacts.First();
-            var newEdition = await HttpRequest.CreateNewEdition(_client, artefact.editionId); // Clone it
+            var newEdition = await EditionHelpers.CreateCopyOfEdition(_client, artefact.editionId); // Clone it
             
             // Act
             var (response, _) = await HttpRequest.SendAsync<string, string>(_client, HttpMethod.Delete,
@@ -313,7 +313,7 @@ namespace SQE.ApiTest
             // Assert
             Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
 
-            await HttpRequest.DeleteEdition(_client, newEdition, true);
+            await EditionHelpers.DeleteEdition(_client, newEdition, true);
         }
         
         /// <summary>
@@ -326,7 +326,7 @@ namespace SQE.ApiTest
             // Arrange
             var allArtefacts = (await GetRandomEditionArtefacts()).artefacts; // Find edition with artefacts
             var artefact = allArtefacts.First();
-            var newEdition = await HttpRequest.CreateNewEdition(_client, artefact.editionId); // Clone it
+            var newEdition = await EditionHelpers.CreateCopyOfEdition(_client, artefact.editionId); // Clone it
             var newArtefactName = _faker.Random.Words(5);
             
             // Act (update name)
@@ -341,7 +341,7 @@ namespace SQE.ApiTest
             // Assert (update name)
             Assert.Equal(HttpStatusCode.Unauthorized, nameResponse.StatusCode);
 
-            await HttpRequest.DeleteEdition(_client, newEdition, true);
+            await EditionHelpers.DeleteEdition(_client, newEdition, true);
         }
         
         /// <summary>
@@ -354,7 +354,7 @@ namespace SQE.ApiTest
             // Arrange
             var allArtefacts = (await GetRandomEditionArtefacts()).artefacts; // Find edition with artefacts
             var artefact = allArtefacts.First();
-            var newEdition = await HttpRequest.CreateNewEdition(_client, artefact.editionId); // Clone it
+            var newEdition = await EditionHelpers.CreateCopyOfEdition(_client, artefact.editionId); // Clone it
             const string newArtefactShape = "POLYGON(0 0,0 200,200 200,0 200,0 0),5 5,5 25,25 25,25 5,5 5),(77 80,77 92,102 92,102 80,77 80))";
             
             // Act (update name)
@@ -370,7 +370,7 @@ namespace SQE.ApiTest
             // Assert (update name)
             Assert.Equal(HttpStatusCode.BadRequest, nameResponse.StatusCode);
 
-            await HttpRequest.DeleteEdition(_client, newEdition, true);
+            await EditionHelpers.DeleteEdition(_client, newEdition, true);
         }
         
         /// <summary>
@@ -383,7 +383,7 @@ namespace SQE.ApiTest
             // Arrange
             var allArtefacts = (await GetRandomEditionArtefacts()).artefacts; // Find edition with artefacts
             var artefact = allArtefacts.First();
-            var newEdition = await HttpRequest.CreateNewEdition(_client, artefact.editionId); // Clone it
+            var newEdition = await EditionHelpers.CreateCopyOfEdition(_client, artefact.editionId); // Clone it
             var newArtefactMatrix = RandomPosition(properlyFormatted: false);
             
             // Act (update name)
@@ -399,7 +399,7 @@ namespace SQE.ApiTest
             // Assert (update name)
             Assert.Equal(HttpStatusCode.BadRequest, nameResponse.StatusCode);
 
-            await HttpRequest.DeleteEdition(_client, newEdition, true);
+            await EditionHelpers.DeleteEdition(_client, newEdition, true);
         }
         #endregion Access artefacts (should fail)
