+++ api-test/Helpers/HttpRequest.cs
@@ -67,7 +67,14 @@ namespace SQE.ApiTest.Helpers
             return (response: response, msg: parsedClass);
         }
         
-        //TODO: these helpers do not belong here.  Move them to a more sensible place.
+        /// <summary>
+        /// Returns the JWT for a user account. If no username/password is provided
+        /// this will return a JWT for the "test" account.
+        /// </summary>
+        /// <param name="client">Http Client</param>
+        /// <param name="username">email address of the user</param>
+        /// <param name="pwd">the user's password</param>
+        /// <returns>A valid JWT</returns>
         public static async Task<string> GetJWTAsync(HttpClient client, string username = null, string pwd = null)
         {
             var name = username ?? "test";
@@ -77,143 +84,5 @@ namespace SQE.ApiTest.Helpers
             response.EnsureSuccessStatusCode();
             return msg.token;
         }
-
-        public static async Task<uint> CreateNewEdition(HttpClient client, uint editionId = 1, string username = null, 
-            string pwd = null)
-        {
-            var newScrollRequest = new EditionUpdateRequestDTO("test-name", null, null);
-            var (response, msg) = await SendAsync<EditionUpdateRequestDTO, EditionDTO>(client, HttpMethod.Post, 
-                $"/v1/editions/{editionId}", newScrollRequest, await GetJWTAsync(client, username, pwd));
-            response.EnsureSuccessStatusCode();
-            return msg.id;
-        }
-        
-        public static async Task DeleteEdition(HttpClient client, uint editionId, bool authenticated = false, bool shouldSucceed = true, string email = null, string pwd = null)
-        {
-            var (response, msg) = await SendAsync<string, DeleteTokenDTO>(client, HttpMethod.Delete, 
-                $"/v1/editions/{editionId}?optional=deleteForAllEditors", null, authenticated ? await GetJWTAsync(client, email, pwd) : null);
-            if (shouldSucceed)
-            {
-                response.EnsureSuccessStatusCode();
-                Assert.NotNull(msg.token);
-                Assert.Equal(editionId, msg.editionId);
-                var (response2, msg2) = await SendAsync<string, DeleteTokenDTO>(client, HttpMethod.Delete, 
-                    $"/v1/editions/{msg.editionId}?optional=deleteForAllEditors&token={msg.token}", null, authenticated ? await GetJWTAsync(client, email, pwd) : null);
-                response2.EnsureSuccessStatusCode();
-                Assert.Null(msg2);
-            }
-        }
-        
-        #region User Account Conveniences
-
-        public static async Task<DetailedUserDTO> CreateRandomUser(HttpClient client, string password)
-        {
-            var faker = new Faker("en");
-            var user = new NewUserRequestDTO(faker.Internet.Email(), password, faker.Company.CompanyName(), 
-                faker.Name.FirstName(), faker.Name.LastName());
-
-            var (userAcctResp, userAcctMsg) = await CreateUserAccountAsync(client, user);
-            userAcctResp.EnsureSuccessStatusCode();
-
-            await ActivateUserAccountAsync(client, userAcctMsg);
-            return userAcctMsg;
-        }
-        
-        private static async Task<(HttpResponseMessage response, DetailedUserDTO msg)> CreateUserAccountAsync(HttpClient client,
-            NewUserRequestDTO user, bool shouldSucceed = true)
-        {
-            var (response, msg) = await HttpRequest.SendAsync<NewUserRequestDTO, DetailedUserDTO>(client, 
-                HttpMethod.Post, "/v1/users", user);
-
-            // Assert
-            if (shouldSucceed)
-            {
-                response.EnsureSuccessStatusCode();
-                Assert.Equal(user.email, msg.email);
-                Assert.Equal(user.forename, msg.forename);
-                Assert.Equal(user.surname, msg.surname);
-                Assert.Equal(user.organization, msg.organization);
-            }
-            else
-            {
-                Assert.Equal(HttpStatusCode.Conflict, response.StatusCode);
-            }
-            return (response, msg);
-        }
-
-        private static async Task ActivateUserAccountAsync(HttpClient client, DetailedUserDTO user, bool shouldSucceed = true)
-        {
-            var userToken = await GetToken(user.email); // Get  token from DB
-            var payload = new AccountActivationRequestDTO() {token = userToken.token};
-            
-            var (response, msg) =
-                await HttpRequest.SendAsync<AccountActivationRequestDTO, UserDTO>(client, HttpMethod.Post, 
-                    "/v1/users/confirm-registration", payload);
-            
-            // Assert
-            if (shouldSucceed)
-            {
-                response.EnsureSuccessStatusCode();
-                Assert.Equal(HttpStatusCode.NoContent, response.StatusCode);
-                var confirmedUser = await GetUserByEmail(user.email);
-                Assert.Equal(user.email, confirmedUser.email);
-                Assert.True(confirmedUser.activated);
-            }
-            else
-            {
-                Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
-            }
-        }
-        
-        private static async Task<UserObj> GetUserByEmail(string email, bool shouldSucceed = true)
-        {
-            var db = new DatabaseQuery();
-            const string checkForUserSQL = "SELECT * FROM user WHERE email = @Email";
-            
-            var checkForUserSQLParams = new DynamicParameters();
-            checkForUserSQLParams.Add("@Email", email);
-            if (shouldSucceed)
-                return await db.RunQuerySingleAsync<UserObj>(checkForUserSQL, checkForUserSQLParams); // Get user from DB
-            
-            var users = await db.RunQueryAsync<UserObj>(checkForUserSQL, checkForUserSQLParams); // Get user from DB
-            Assert.Empty(users);
-            return new UserObj();
-        }
-        
-        private static async Task<Token> GetToken(string email, bool shouldSucceed = true)
-        {
-            var db = new DatabaseQuery();
-            const string getNewUserTokenSQL = "SELECT token, date_created, type FROM user_email_token JOIN user USING(user_id) WHERE email = @Email";
-            
-            var checkForUserSQLParams = new DynamicParameters();
-            checkForUserSQLParams.Add("@Email", email);
-            
-            if (shouldSucceed)
-                return await db.RunQuerySingleAsync<Token>(getNewUserTokenSQL, checkForUserSQLParams); // Get  token from DB
-            
-            var tokens = await db.RunQueryAsync<Token>(getNewUserTokenSQL, checkForUserSQLParams);
-            Assert.Empty(tokens);
-            return new Token();
-        }
-        
-        private class UserObj
-        {
-            public string email { get; set; }
-            public string password { get; set; }
-            public string forename { get; set; }
-            public string surname { get; set; }
-            public string organization { get; set; }
-            public bool activated { get; set; }
-            public uint user_id { get; set; }
-        }
-        
-        private class Token
-        {
-            public string token { get; set; }
-            public string type { get; set; }
-            public DateTime date_created { get; set; }
-        }
-        
-        #endregion User Account Conveniences
     }
 }
